Move Instruction to App\Slave
