+++ src/Command/ProbeDispatcherCommand.php
@@ -8,8 +8,8 @@ use App\DependencyInjection\Queue;
 use App\DependencyInjection\SlaveConfiguration;
 use App\DependencyInjection\StatsManager;
 use App\DependencyInjection\WorkerManager;
-use App\Instruction\Instruction;
 use App\Process\SymfonyProcessFactory;
+use App\Slave\Instruction;
 use App\Slave\Task\FetchConfiguration;
 use App\Slave\Task\PublishStatistics;
 use App\Version\GitVersionReader;
